+++ HISTORY.rst
@@ -9,6 +9,7 @@
 
 - 1173_: introduced PSUTIL_DEBUG environment variable which can be set in order
   to print useful debug messages on stderr (useful in case of nasty errors).
+- 1177_: added support for sensors_battery() on OSX.  (patch by Arnon Yaari)
 
 **Bug fixes**