+++ HISTORY.rst
@@ -17,6 +17,7 @@
 - 1169_: [Linux] users() "hostname" returns username instead.  (patch by
   janderbrain)
 - 1172_: [Windows] `make test` does not work.
+- 1181_: [OSX] Process.memory_maps() may raise ENOENT.
 
 5.4.1
 =====